Rename navbar examples file
